fix: align timeout type annotations from int to float for consistency
The Command.timeout, DEFAULT_COMMAND_TIMEOUT, and run_command timeout
parameter were typed as int while RunConfig.timeout and agent timeout
used float | None. Since subprocess.run() accepts float and the CLI
already validates isinstance(timeout, (int, float)), this aligns the
types to match actual usage.

+++ src/ralphify/_run_types.py
@@ -15,7 +15,7 @@ from enum import Enum
 from pathlib import Path
 
 
-DEFAULT_COMMAND_TIMEOUT: int = 60
+DEFAULT_COMMAND_TIMEOUT: float = 60
 """Default timeout in seconds for commands defined in RALPH.md frontmatter."""
 
 RUN_ID_LENGTH: int = 12
@@ -54,7 +54,7 @@ class Command:
 
     name: str
     run: str
-    timeout: int = DEFAULT_COMMAND_TIMEOUT
+    timeout: float = DEFAULT_COMMAND_TIMEOUT
 
 
 @dataclass
